canvas is exported to post draw session and can trigger a download
